[PLT-3191] Ability to fetch secrets from secrets manager
This introduces a `fetch` class, that given a list of secrets and desired versions, will attempt to bulk fetch those secrets. Any secrets it cannot locate at the latest version, it will fall back to looking up one-by-one at the specific version requested.

This adds a method to the base module that enables loading all secret values into ENV.

This method is what will be called by the users of the gem to do the thing.